reorganized packages. centralized dependencies.
