+++ app/build.gradle
@@ -1,4 +1,7 @@
 apply plugin: 'com.android.application'
+apply plugin: 'com.neenbedankt.android-apt'
+
+def globalConf = rootProject.ext
 
 android {
     compileSdkVersion 23
@@ -20,8 +23,19 @@ android {
 }
 
 dependencies {
+    Map<String, String> dependencies = globalConf.dependencies
+
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:23.4.0'
+    testCompile dependencies.junit
+    compile dependencies.appCompat
+
+    // TODO - remove this
     compile 'commons-io:commons-io:2.4'
+
+    // squidb
+    compile dependencies.squidb
+    compile dependencies.squidbAnnotation
+    apt dependencies.squidbProcessor
+    compile dependencies.squidbRecyclerView
+    compile dependencies.squidbSupportLoader
 }
