+++ app/src/main/java/com/codepath/simpletodo/activities/MainActivity.java
@@ -1,4 +1,4 @@
-package com.codepath.simpletodo;
+package com.codepath.simpletodo.activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -9,6 +9,8 @@ import android.widget.ArrayAdapter;
 import android.widget.EditText;
 import android.widget.ListView;
 import android.widget.Toast;
+
+import com.codepath.simpletodo.R;
 
 import org.apache.commons.io.FileUtils;
